pyproject.toml
[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[project]
name = "exif-extractor"
version = "0.1.0"
description = "An MCP server built to extract EXIF information from images"
authors = [
    {name = "Yujia Qian", email = "yjqian19@gmail.com"}
]

requires-python = ">=3.10"
dependencies = [
    "mcp>=1.15.0",
    "smithery>=0.4.2",
    "pillow>=10.0.0",
    "piexif>=1.1.3",
    "requests>=2.31.0",
]

[project.scripts]
dev = "smithery.cli.dev:main" # Run the MCP server in development mode
start = "smithery.cli.start:main" # Run the MCP server in production mode
playground = "smithery.cli.playground:main" # Run server with interactive testing playground

[tool.smithery]
server = "exif_extractor.server:create_server" # Function that returns a patched FastMCP server
log_level = "warning" # Production log level (optional, defaults to "warning")
